+++ spuce/filters/raised_cosine_imp.h
@@ -4,6 +4,6 @@
 
 #include <spuce/typedefs.h>
 namespace spuce {
-float_type raised_cosine_imp(float_type alpha, float_type xin, float_type rate, long num_taps);
+float_type raised_cosine_imp(float_type alpha, float_type xin, float_type rate, size_t num_taps);
 
 }  // namespace spuce
